LayoutTests/fast/js/dfg-convert-this-other-then-exit-on-object-expected.txt

Tests that a convert_this optimized for the case where this is null behaves correctly when you then pass an object.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS foo.call(me) is 42
PASS foo.call(me) is 42
PASS foo.call(me) is 42
PASS foo.call(me) is 42
PASS foo.call(me) is 42
PASS foo.call(me) is 42
PASS foo.call(me) is 42
PASS foo.call(me) is 42
PASS foo.call(me) is 42
PASS foo.call(me) is 42
PASS foo.call(me) is 42
PASS foo.call(me) is 42
PASS foo.call(me) is 42
PASS foo.call(me) is 42
PASS foo.call(me) is 42
PASS foo.call(me) is 42
PASS foo.call(me) is 42
PASS foo.call(me) is 42
PASS foo.call(me) is 42
PASS foo.call(me) is 42
PASS foo.call(me) is 42
PASS foo.call(me) is 42
PASS foo.call(me) is 42
PASS foo.call(me) is 42
PASS foo.call(me) is 42
PASS foo.call(me) is 42
PASS foo.call(me) is 42
PASS foo.call(me) is 42
PASS foo.call(me) is 42
PASS foo.call(me) is 42
PASS foo.call(me) is 42
PASS foo.call(me) is 42
PASS foo.call(me) is 42
PASS foo.call(me) is 42
PASS foo.call(me) is 42
PASS foo.call(me) is 42
PASS foo.call(me) is 42
PASS foo.call(me) is 42
PASS foo.call(me) is 42
PASS foo.call(me) is 42
PASS foo.call(me) is 42
PASS foo.call(me) is 42
PASS foo.call(me) is 42
PASS foo.call(me) is 42
PASS foo.call(me) is 42
PASS foo.call(me) is 42
PASS foo.call(me) is 42
PASS foo.call(me) is 42
PASS foo.call(me) is 42
PASS foo.call(me) is 42
PASS foo.call(me) is 42
PASS foo.call(me) is 42
PASS foo.call(me) is 42
PASS foo.call(me) is 42
PASS foo.call(me) is 42
PASS foo.call(me) is 42
PASS foo.call(me) is 42
PASS foo.call(me) is 42
PASS foo.call(me) is 42
PASS foo.call(me) is 42
PASS foo.call(me) is 42
PASS foo.call(me) is 42
PASS foo.call(me) is 42
PASS foo.call(me) is 42
PASS foo.call(me) is 42
PASS foo.call(me) is 42
PASS foo.call(me) is 42
PASS foo.call(me) is 42
PASS foo.call(me) is 42
PASS foo.call(me) is 42
PASS foo.call(me) is 42
PASS foo.call(me) is 42
PASS foo.call(me) is 42
PASS foo.call(me) is 42
PASS foo.call(me) is 42
PASS foo.call(me) is 42
PASS foo.call(me) is 42
PASS foo.call(me) is 42
PASS foo.call(me) is 42
PASS foo.call(me) is 42
PASS foo.call(me) is 42
PASS foo.call(me) is 42
PASS foo.call(me) is 42
PASS foo.call(me) is 42
PASS foo.call(me) is 42
PASS foo.call(me) is 42
PASS foo.call(me) is 42
PASS foo.call(me) is 42
PASS foo.call(me) is 42
PASS foo.call(me) is 42
PASS foo.call(me) is 42
PASS foo.call(me) is 42
PASS foo.call(me) is 42
PASS foo.call(me) is 42
PASS foo.call(me) is 42
PASS foo.call(me) is 42
PASS foo.call(me) is 42
PASS foo.call(me) is 42
PASS foo.call(me) is 42
PASS foo.call(me) is 42
PASS foo.call(me) is 42
PASS foo.call(me) is 42
PASS foo.call(me) is 42
PASS foo.call(me) is 42
PASS foo.call(me) is 42
PASS foo.call(me) is 42
PASS foo.call(me) is 42
PASS foo.call(me) is 42
PASS foo.call(me) is 42
PASS foo.call(me) is 42
PASS foo.call(me) is 42
PASS foo.call(me) is 42
PASS foo.call(me) is 42
PASS foo.call(me) is 42
PASS foo.call(me) is 42
PASS foo.call(me) is 42
PASS foo.call(me) is 42
PASS foo.call(me) is 42
PASS foo.call(me) is 42
PASS foo.call(me) is 42
PASS foo.call(me) is 42
PASS foo.call(me) is 42
PASS foo.call(me) is 42
PASS foo.call(me) is 42
PASS foo.call(me) is 42
PASS foo.call(me) is 42
PASS foo.call(me) is 42
PASS foo.call(me) is 42
PASS foo.call(me) is 42
PASS foo.call(me) is 42
PASS foo.call(me) is 42
PASS foo.call(me) is 42
PASS foo.call(me) is 42
PASS foo.call(me) is 42
PASS foo.call(me) is 42
PASS foo.call(me) is 42
PASS foo.call(me) is 42
PASS foo.call(me) is 42
PASS foo.call(me) is 42
PASS foo.call(me) is 42
PASS foo.call(me) is 42
PASS foo.call(me) is 42
PASS foo.call(me) is 42
PASS foo.call(me) is 42
PASS foo.call(me) is 42
PASS foo.call(me) is 42
PASS foo.call(me) is 42
PASS foo.call(me) is 42
PASS foo.call(me) is 42
PASS foo.call(me) is 42
PASS foo.call(me) is 42
PASS foo.call(me) is 42
PASS foo.call(me) is 42
PASS foo.call(me) is 42
PASS foo.call(me) is 42
PASS foo.call(me) is 42
PASS foo.call(me) is 42
PASS foo.call(me) is 42
PASS foo.call(me) is 42
PASS foo.call(me) is 42
PASS foo.call(me) is 42
PASS foo.call(me) is 42
PASS foo.call(me) is 42
PASS foo.call(me) is 42
PASS foo.call(me) is 42
PASS foo.call(me) is 42
PASS foo.call(me) is 42
PASS foo.call(me) is 42
PASS foo.call(me) is 42
PASS foo.call(me) is 42
PASS foo.call(me) is 42
PASS foo.call(me) is 42
PASS foo.call(me) is 42
PASS foo.call(me) is 42
PASS foo.call(me) is 42
PASS foo.call(me) is 42
PASS foo.call(me) is 42
PASS foo.call(me) is 42
PASS foo.call(me) is 42
PASS foo.call(me) is 42
PASS foo.call(me) is 42
PASS foo.call(me) is 42
PASS foo.call(me) is 42
PASS foo.call(me) is 42
PASS foo.call(me) is 42
PASS foo.call(me) is 42
PASS foo.call(me) is 42
PASS foo.call(me) is 42
PASS foo.call(me) is 42
PASS foo.call(me) is 42
PASS foo.call(me) is 42
PASS foo.call(me) is 42
PASS foo.call(me) is 42
PASS foo.call(me) is 42
PASS foo.call(me) is 42
PASS foo.call(me) is 42
PASS foo.call(me) is 42
PASS foo.call(me) is 42
PASS foo.call(me) is 42
PASS foo.call(me) is 42
PASS successfullyParsed is true

TEST COMPLETE
